Experimental query response filter mechanism (#3)
This PR adds the experimental endpoint extract_rate to the canister.
When providing a query response and a (jq) filter, it extracts the rate and returns the rate as an unsigned integer scaled by a factor of 100.
Note that the PR uses the jaq (core) library instead of jq, which apparently offers greater performance. Moreover, it does not depend on (the C library) jqlib, which is not entirely trivial to set up correctly.
A simple example test:
Query response (first parameter): {"name":"apple","color":"green","price":1.20}
Filter (second parameter): .price
The result should be 120.