Bug fix: jaq filters that include timestamps in milliseconds fail compilation inside of the canister. (#19)
When running the `xrc` canister through the upcoming e2e tests, it was
noticed that Binance and Okx were failing to retrieve their rates. The
`dfx` log file showed the following error:

```
    Extract {
        exchange: "Binance",
        error: MalformedFilterExpression {
            filter: "map(select(.[0] == 1614596340000))[0][1] | tonumber",
            errors: [
                "found end of input",
            ],
        },
    }
```

The error is caused by `jaq` trying to box the timestamp into a `u32`.
The solution is to compare the timestamps as strings for these
exchanges.