Introducing an exchange rate cache (#25)
This PR introduces a caching mechanism for cryptocurrency rates.
It currently supports only insertion and retrieval. Moreover, it already
implements an automatic clean-up mechanism, removing expired rates.
However, the implemented mechanism is not good enough because the cache
may be filled up with many different rates in a short amount of time.

A follow-up PR will introduce an LRU-based pruning mechanism. Some
pieces of code are already there for this mechanism (for example, the
logical timestamps). These pieces don't make much sense now, so take
that into consideration when reviewing the PR.

Further note that the cache is not used in `main.rs` yet. I added
`#[allow(dead_code)]` to silence annoying output when compiling. When
the cache functionality is integrated, these directives must be removed.