Adding IPv6 and fiat currency support information to the exchanges (#37)
This PR adds the functions

- `supports_ipv6`
- `suppported_fiat_currencies`

to the exchanges.

The first function is needed to determine whether a NAT64 router must be
used in the HTTPS calls and the second function is used to request the
fiat currency as the quote asset directly, rather than going through a
conversion process.

Note that there are some caveats:
- Binance technically supports EUR but it is limited to a small number
of cryptocurrencies. Moreover, Binance only supports cryptocurrency/fiat
currency pairs UNLESS the cryptocurrency is a stablecoin, in which case
it only supports fiat currency/cryptocurrency. Since it would be too
cumbersome to implement these rules, I decided to ignore the EUR support
for Binance.
- Coinbase supports USD, EUR, and GBP; however, the number of
cryptocurrency/fiat currency pairs depends on the fiat currency (there
are many for USD, less for EUR, and even less for GBP). For the sake of
simplicity, the `suppported_fiat_currencies` function simply declares
that Coinbase supports these fiat currencies, which means that the call
might fail if somebody tries to get the rate for some lesser known
cryptocurrency. I think this is fine because other exchanges should
still return rates (with respect to USDT).