`e2e-tests`: Route calls by path and provided parameters in the query string. (#41)
The following PR updates the e2e tests so that the mock cryptoexchanges
can respond differently based on the request completely. A Lua-based
module is added to the nginx instance that transforms the request URI to
a file path to where the expected JSON is located in the container.
Prior to this change all requests would point to a singular JSON per
mock exchange.