Introducing the `QueriedExchangeRate` (#43)
This PR introduces a separation between `ExchangeRate` and
`QueriedExchangeRate`. The former is used to encode results for
requests, the latter is used internally.
The reason for this separation is that it is not possible to cleanly
carry out operations on the `ExchangeRate` struct (e.g., `A/B * B/C =
A/C` or `A/B / C/B = A/C`). The main problem is that there is not enough
information to compute the standard deviation.

If this PR is merged, necessary follow-up work is to change `api.rs` to
use the `QueriedExchangeRate` as well, removing the division operator
from `ExchangeRate`. The idea is to work exclusively with
`QueriedExchangeRate` until an `ExchangeRate` struct is needed for the
response, which already supports the conversion from a
`QueriedExchangeRate` struct.