`forex`: Implement `handle_fiat_pair` and update the `ForexRateStore::get` (#64)
This PR updates `ForexRateStore::get` to return a `Result<Forex,
GetForexRateError>` type allowing the reasons why a forex rate could not
be found to be exposed. It then implements the `ForexRateStore` in
`handle_fiat_pair`.

A follow-up PR will be needed to e2e test this. It is not in this PR as
we still need to implement the periodic update for collecting rates.

Co-authored-by: Thomas Locher <thomas.locher@dfinity.org>