`refactor`: `ExchangeRateCache::insert` should simply panic when the quote asset does not match the expected quote asset. (#78)
This PR removes the `Result` from the `ExchangeRateCache::insert`. We
are currently using the `Result` to panic when the cache is being
misused. `insert` does not issue any other errors. To clean up the
`expect`s from the code, it would be simpler to use `panic!` in the one
location that we need it, inside `insert`.