Checking rate validity (#95)
As requested by product security, this PR introduces an exchange rate
validity check.

Concretely, it checks that there are `n/2+1` rates out of `n` collected
rates with bounded deviation. The deviation can be at most 10% of the
smallest rate in the set of `n/2+1` rates. Naturally, the threshold of
10% can be adjusted if necessary.

Note that the rates in a `QueriedExchangeRate` struct must be sorted for
the validity check to work correctly. The rates are now automatically
sorted when using the constructor or when performing any of the
supported operations on a `QueriedExchangeRate` instance. However, care
has to be taken when instantiating an instance without the use of `new`.