Fixing the rate validity check (#105)
The logic of the validity check was wrong: Instead of checking that
there is a majority of rates that are close, the code verified that
every majority of rates has a small deviation. This is clearly wrong as
a single bad rate could make the whole set invalid.
This bug is probably also the reason why the rates in the e2e tests were
not allowed to deviate much without causing a failure.

This PR fixes the validity check. It also extends the validity unit test
to make sure that a single, arbitrarily small or large rate cannot cause
the rate to be considered invalid by itself.