`refactor`: Remove `time_secs` call from `ForexRateStore::get`. (#111)
This PR removes the `time_secs` call from `ForexRateStore::get`. The
call causes the unit tests when running them on M1 Macs. This is
resolved by adding a parameter to `get` to intake the current timestamp.
This allows a timestamp to be provided by an `Environment`
implementation in the unit tests.