Log all errors that can be encountered when making outbound HTTP calls. (#126)
This PR reverts a few logging statements in the transforms in favor of
logging near the calling code. This allows us to capture all errors in
the logs when making outbound HTTP calls. There will be a follow up PR
to clean up the logging further and use the various environments.