`cryptoexchanges`: Add `max_response_bytes` for cryptoexchanges (#121)
This PR sets the `max_response_bytes` for HTTP outcalls to
cryptoexchanges. It uses `500` as a default as most exchanges provide
responses at 100-200 bytes. This number can be tweaked per exchange, but
setting it 500 provides some safety. In another PR, we can add tracking
for responses that go over this limit.