Changing the behavior when no timestamp is provided in a request (#156)
Currently, if a request doesn't specify a timestamp, the start of the
most recent minute is used as the timestamp.
Since it takes some time for exchanges to offer data for the current
minute, this approach results in many failed calls.

This PR changes the behavior: The start of the current minute is only
used if at least 30 seconds have passed. Otherwise, the start of the
previous minute is used.