feat: Allow e2e-tests to take arguments so one test can be ran easily (#166)
This PR adds the ability to `./scripts/e2e-tests` to run a singular
test. During development of new `e2e` tests, it is generally painful to
run a singular test. This is addressed by giving the option to add a
flag (`n`) to provide a name of the test that a user wants to run.

Example: `./scripts/e2e-tests -n test_name_here`