feat: Add `disable-forex-weekend-check` feature flag. (#170)
This PR adds the `disable-forex-weekend-check` feature flag. This flag
is used to skip the check to see if a timestamp is on a Saturday or
Sunday when attempting to contact forex sources. This is being
introduced in order to add forex sources to the e2e tests. If the e2e
tests were to be run on a Saturday or Sunday, the e2e tests using the
forex sources mock data would fail.