`bugfix`: fiat-crypto pairs were not handled correctly in earlier refactor (#175)
When deploying the timestamp = null feature, ForexQuoteAssetNotFound
errors started to populate the logs. This lead to the discovery that the
refactor missed the inversion of the base and quote assets for
fiat-crypto pairs.

This PR addresses the issue by inverting the base and quote assets for a
fiat-crypto pair. The inverted request is then passed in to the
`handle_crypto_base_fiat_quote_pair` handle function. To ensure coverage
to avoid this problem from occurring in the future, additional unit
tests were added to ensure this case is covered.