Updating the OKX logic to set the start time 1 minute in the past (#181)
OKX never returns data for the current minute, which is problematic when
no timestamp is specified in the request.

This PR proposes and implements a simple fix: A minute is subtracted
from the start time. This change has the following effect:

- If no timestamp is set in the request, it will provide the latest
available data, that is, the rate from the previous minute.
- If a timestamp (in the past) is set in the request, the response may
contain two data points. Since the more recent data is provided at index
0, the correct data is extracted.

Co-authored-by: Ryan Croote <ryan.croote@dfinity.org>