`bugfix`: Central Bank of Bosnia Herzegovina added rates on a Saturday (#200)
This PR fixes a bug that caused the Central Bank of Bosnia Herzegovina
to be added on a Saturday. The bug came from the query timestamp (B.H.
needs to use +1 day to get the rates for the prior day) being returned
as the timestamp that the rates should be assigned to.