fix: In order to ensure consensus passes, forex rates must be sorted. (#247)
As of writing this description, the XRC on production has only been able
to collect rates from 2 forex sources due to consensus errors. This is
due to encoding a hash map to candid. The hash map does not preserve any
sense of order; therefore, a b-tree map should be used as it will
preserve order of the values when encoded.

This PR also adds logging into the transform so we may better see what
is actually occurring during that process.