fix: Remove access by index when inverting and dividing rates (#257)
If no rates are retained while instantiating a `QueriedExchangeRate`,
the inner `rates` field can become empty. This causes a panic when
inverting and dividing rates. This PR patches this behavior to allow the
logic to reach the inevitable `validate` calls the occur in each
`handle_*_pair` function in `api.rs`.

This is not a major bug as we have implemented `Drop` for our various
guards, but it is still a good idea to remove panics.