docs: Explain exchange rate terminology. (#270)
In particular, what does "X/Y" mean? What are "base" and "quote" assets?

In short in "X/Y", X is the "base" asset, and Y is the "quote" asset,
and it tells you the price of X in terms of Y.

At least for me, this is really not self-explanatory.